Refactored the app to common wave app style
